review comment, and move helper function to services
