+++ lib/services/database_test.go
@@ -26,6 +26,8 @@ import (
 	"github.com/gravitational/teleport/lib/defaults"
 	"github.com/gravitational/teleport/lib/fixtures"
 	"github.com/gravitational/teleport/lib/utils"
+	"github.com/pborman/uuid"
+	"github.com/sirupsen/logrus"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/rds"
@@ -188,4 +190,42 @@ func TestDatabaseFromRDSCluster(t *testing.T) {
 	actual, err := NewDatabaseFromRDSCluster(cluster)
 	require.NoError(t, err)
 	require.Equal(t, expected, actual)
+}
+
+func TestAuroraMySQLVersion(t *testing.T) {
+	require.Equal(t, "5.6.10a", auroraMySQLVersion(&rds.DBCluster{EngineVersion: aws.String("5.6.10a")}))
+	require.Equal(t, "1.22.1", auroraMySQLVersion(&rds.DBCluster{EngineVersion: aws.String("5.6.mysql_aurora.1.22.1")}))
+	require.Equal(t, "1.22.1.3", auroraMySQLVersion(&rds.DBCluster{EngineVersion: aws.String("5.6.mysql_aurora.1.22.1.3")}))
+}
+
+func TestIsRDSClusterSupported(t *testing.T) {
+	logger := logrus.WithField("test", t.Name())
+
+	tests := []struct {
+		name          string
+		engineMode    string
+		engineVersion string
+		isSupported   bool
+	}{
+		{name: "provisioned", engineMode: RDSEngineModeProvisioned, engineVersion: "5.6.mysql_aurora.1.22.0", isSupported: true},
+		{name: "serverless", engineMode: RDSEngineModeServerless, engineVersion: "5.6.mysql_aurora.1.22.0", isSupported: false},
+		{name: "parallel query supported", engineMode: RDSEngineModeParallelQuery, engineVersion: "5.6.mysql_aurora.1.22.0", isSupported: true},
+		{name: "parallel query unsupported", engineMode: RDSEngineModeParallelQuery, engineVersion: "5.6.mysql_aurora.1.19.6", isSupported: false},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			cluster := &rds.DBCluster{
+				DBClusterArn:        aws.String("arn:aws:rds:us-east-1:1234567890:cluster:test"),
+				DBClusterIdentifier: aws.String(test.name),
+				DbClusterResourceId: aws.String(uuid.New()),
+				Engine:              aws.String(RDSEngineAuroraMySQL),
+				EngineMode:          aws.String(test.engineMode),
+				EngineVersion:       aws.String(test.engineVersion),
+			}
+
+			require.Equal(t, test.isSupported, IsRDSClusterSupported(cluster, logger))
+
+		})
+	}
 }
